+++ lib/roo_on_rails/checks/environment.rb
@@ -6,7 +6,6 @@ require 'roo_on_rails/checks/heroku/app_exists'
 require 'roo_on_rails/checks/sidekiq/settings'
 require 'roo_on_rails/checks/heroku/drains_metrics'
 require 'roo_on_rails/checks/documentation/playbook'
-require 'roo_on_rails/checks/google_oauth/initializer'
 require 'roo_on_rails/checks/papertrail/all'
 
 module RooOnRails
@@ -17,8 +16,16 @@ module RooOnRails
       requires Heroku::PrebootEnabled
       requires Sidekiq::Settings
       requires Documentation::Playbook
-      requires GoogleOauth::Initializer
       requires Papertrail::All
+
+      def initialize(options = {})
+        super
+        @dry_run = options.fetch(:dry_run, false)
+      end
+
+      def run
+        super unless @dry_run
+      end
 
       def call
         # nothing to do